+++ src/classic/spp_server.c
@@ -126,7 +126,7 @@ static void spp_create_sdp_record_internal(uint8_t *service, uint32_t service_re
 	
 	// 0x0100 "ServiceName"
 	de_add_number(service,  DE_UINT, DE_SIZE_16, 0x0100);
-	de_add_data(service,  DE_STRING, strlen(name), (uint8_t *) name);
+	de_add_data(service,  DE_STRING, (uint16_t) strlen(name), (uint8_t *) name);
 }
 
 void spp_create_sdp_record(uint8_t *service, uint32_t service_record_handle, int rfcomm_channel, const char *name){
